Move HF Ref model out of grpo into ref_model file
